% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printDET}
\alias{printDET}
\title{DET Curve information}
\usage{
printDET(dets)
}
\arguments{
\item{dets}{list of DET curves originated by detc() or detc.CI() functions}
}
\description{
From a DET curve list (generated with detc() or detc.CI()), the function shows the different atributtes of each curve with a little description.
}
