% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,DETs-method}
\alias{show,DETs-method}
\title{Show the structure of a DETs object}
\usage{
\S4method{show}{DETs}(object)
}
\arguments{
\item{object}{An object of class \code{DETs}.}
}
\description{
Registers the S4 \code{show} method for \code{DETs} objects so that
\code{show(object)} and automatic console display use the same output as
\code{show.DETs(object)}.
}
