% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show.DETs}
\alias{show.DETs}
\title{Show the structure of a DET object}
\usage{
show.DETs(dets)
}
\arguments{
\item{dets}{An object of class "DETs".}
}
\description{
From a 'DETs' object (generated with the \code{detc} function), the function shows the different attributes of each curve with a brief description, which have to be used to get the results for each curve.
}
