% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detc.R
\name{detc}
\alias{detc}
\title{DET Curve calculation}
\usage{
detc(
  response = NULL,
  predictors = NULL,
  dets = NULL,
  names = NULL,
  conf = NULL,
  positive = "",
  parallel = FALSE,
  ncores = detectCores(),
  nboot = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{response}{A factor, typically encoded with 0 (healthy, genuine user, signal, normal) and 1 (diseased, impostor user, noise, abnormal). It can also be a dichotomous variable that will be treated as a factor. By default, the second factor level is used as the positive class.}

\item{predictors}{A matrix whose columns represent the values of each predictor.}

\item{dets}{A 'DETs' object which will be used to compute the DET curves.}

\item{names}{A character vector that will be used to set the names of the DET curves. They will also appear in the plot legend.}

\item{conf}{If present, the confidence level for the DET curve CI, within [0,1]. Default: the CI is not computed.}

\item{positive}{A string with the name of the 'positive' level used as the reference level of 'response'. If left as the default empty string, the second factor level of 'response' is used.}

\item{parallel}{If TRUE, the bootstrap method used to calculate the CI is processed in parallel, using the backend provided by \code{plyr} (foreach).}

\item{ncores}{The number of nodes to be forked for the parallel computation of the CI. Default: the maximum available. None used if \code{parallel = FALSE}.}

\item{nboot}{The number of bootstrap replicates to be used for the computation of the CI. Default: 2000.}

\item{plot}{If TRUE, the DET curves are plotted. Default: FALSE.}

\item{...}{Further attributes that will be passed to the \code{plot} function.}
}
\value{
A 'DETs' object. The object's 'detCurves' attribute contains one DET curve per classifier. Each DET curve is an object of class "DET", containing the DET curve parameters (false positive ratio, false negative ratio, and thresholds), along with the Equal Error Rate (EER). If the CI was calculated, the object also includes the median,
upper, and lower CI limits for the false negative ratio and EER. Use the \code{show} method for more details about the results saved in a 'DETs' object.
}
\description{
From a response and one or more predictors, the function calculates a DET curve for each response-predictor pair. Optionally, it can compute each curve with a confidence interval (CI).
Instead of a response and predictors, it can also receive a 'DETs' object to extract existing DET curve results and compute CIs.
}
\examples{
\donttest{
library(DET)
n = 500
# Predictors with normal distribution
set.seed(1235)
scoreNegative1 = rnorm(n, mean = 0.25, sd = 0.125)
set.seed(5321)
scoreNegative2 = rnorm(n, mean = 0.25, sd = 0.125)
set.seed(11452)
scorePositive1 = rnorm(n, mean = 0.55, sd = 0.125)
set.seed(54321)
scorePositive2 = rnorm(n, mean = 0.65, sd = 0.125)
response = as.factor(c(rep(c("diseased"), times = n), rep(c("healthy"), times = n)))
predictor1 = c(scoreNegative1, scorePositive1)
predictor2 = c(scoreNegative2, scorePositive2)
predictors = matrix(c(predictor1, predictor2), ncol = 2)
colnames(predictors) = c("DET1", "DET2")
detCurves = detc(
  response,
  predictors,
  positive = "diseased",
  names = colnames(predictors)
)

# Run in parallel for faster execution by activating the logical argument
# 'parallel' and setting the number of cores of your computer
detCurvesWithConfidenceInterval = detc(
  response,
  predictors,
  positive = "diseased",
  names = colnames(predictors),
  conf = 0.95,
  parallel = TRUE,
  ncores = 2
)
}
}
