% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{minDcf}
\alias{minDcf}
\title{Minimum of the Detection Cost Function computation}
\usage{
minDcf(det, p = 0.01, cFp = 1, cFn = 10)
}
\arguments{
\item{det}{An object of class "DET".}

\item{p}{A single numeric value into the (0, 1) intervalrepresenting the prior probability of positive class.}

\item{cFp}{A single numeric value representing the cost of False Positives.}

\item{cFn}{A single numeric value representing the cost of False Negatives.}
}
\value{
A 'data.frame' with two attributes:

- 'minDcfValue': the computed minDCF.

- 'minDcfIndex': the index of the fpr and fnr in which the minimum is reached.
}
\description{
From a 'DET' object, the function computes the minimum value of the Detection Cost Function (minDCF).
}
